# spec/openapi.yaml
openapi: 3.1.0
info:
  title: ADP API
  version: 0.3.0
  description: >
    Native and compat-plane HTTP surface for ADP — see docs/pragmatic_mvp.md
    §2.4 for what is in scope and §2.5 for what is deliberately cut.

    Two planes share this file. `/api/v3` is the GitHub-compatible surface, kept
    faithful enough that unmodified `gh` works against it. `/api/adp` is the
    native plane, which has no GitHub analogue: workspaces, candidate sets,
    sessions and checkpoints, the operation log, and evidence bundles.

    Completeness is enforced, not aspirational: server/src/spec-coverage.test.ts
    enumerates the server's real route table and fails if this file and the
    implementation describe different APIs in either direction. Deliberate
    omissions are the git smart-HTTP wire protocol (specified by git itself and
    proxied verbatim), the loopback-only receive-path hook endpoints, and the
    inbound GitHub webhook receiver, whose payload contract is GitHub's.

    `info.version` above is the contract version, and every response carries it
    in an `ADP-API-Version` header — including 401s and 404s, so a generated
    client can pin before it authenticates. The git smart-HTTP routes are the
    exception: they are proxied verbatim and are not part of this contract.
    See docs/api-compatibility.md for what a version bump promises.
servers:
  - url: /
paths:
...
  /api/adp/repos/{owner}/{repo}/gate-jobs:
    parameters:
      - $ref: "#/components/parameters/owner"
      - $ref: "#/components/parameters/repo"
    get:
      parameters:
        - name: limit
          in: query
          schema: { type: integer, maximum: 200, default: 50 }
        - name: cursor
          in: query
          description: Opaque keyset cursor from a previous page's ADP-Next-Cursor response header.
          schema: { type: string }
      x-required-scope: "repo:read"
      summary: List gate jobs for this repo, newest first.
      security:
        - bearerAuth: []
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "200":
          description: Gate jobs
          content:
            application/json:
              schema:
                type: object
                required: [gate_jobs]
                properties:
                  gate_jobs:
                    type: array
                    items: { $ref: "#/components/schemas/GateJob" }
    post:
      x-required-scope: "repo:write"
      summary: >
        M4-9a: enqueue an isolated gate run. Queue only — no runner executes
        anything here yet (M4-9b); a runner instance-wide polls
        `/api/adp/gate-jobs/claim` and reports back via
        `/api/adp/gate-jobs/{id}/complete`.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [git_sha, name, image, command]
              properties:
                git_sha:
                  type: string
                  pattern: "^[0-9a-f]{40}$"
                name:
                  type: string
                image:
                  type: string
                command:
                  type: string
                timeout_ms:
                  type: integer
                  default: 300000
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "201":
          description: Queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GateJob"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Repository not found
...
  /api/adp/gate-jobs/claim:
    post:
      x-required-scope: "runner"
      summary: >
        Instance-wide: claim the oldest queued gate job, atomically (`FOR
        UPDATE SKIP LOCKED`), regardless of which repo it belongs to. Requires
        the `runner` scope, distinct from `repo:write`/`repo:read` — a runner
        process serves the whole instance and must not hold repo-scoped
        write access just to pull work.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [claimed_by]
              properties:
                claimed_by:
                  type: string
                  description: Opaque runner identifier (hostname, pid, etc.) — never branched on.
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "200":
          description: A job was claimed and is now `running`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GateJob"
        "204":
          description: No queued jobs
  /api/adp/gate-jobs/{id}/checkout:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      x-required-scope: "runner"
      summary: >
        M4-9c: a tarball (`git archive --format=tar`) of the job's own
        repo/sha — "materializes a checkout" (pragmatic_mvp.md's gate-runner
        design) for the isolated executor to `docker cp` into a container,
        never a host mount and never routed through the container's own
        (network-denied) network. Requires the `runner` scope. Bounded to
        jobs currently `running` and served only to the identity whose claim
        holds the job: the window in which this route grants any read access
        to a repo's contents at all is the same narrow window the caller
        itself is actively working the job in.
      security:
        - bearerAuth: []
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "200":
          description: A tar archive of the tree at the job's git_sha
          content:
            application/x-tar:
              schema:
                type: string
                format: binary
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: The job is held by a different runner identity's claim
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Gate job not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Gate job is not `running` — not currently claimed
  /api/adp/gate-jobs/{id}/complete:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      x-required-scope: "runner"
      summary: >
        Report the result of a claimed gate job. Requires the `runner` scope.
        Only a job currently `running` can be completed, and only by the
        identity whose claim holds it — completion writes signed gate
        evidence that land policy trusts.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [succeeded, failed, timed_out, error]
                exit_code:
                  type: integer
                logs:
                  type: string
      responses:
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Missing or invalid token
        "200":
          description: Recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GateJob"
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: The job is held by a different runner identity's claim
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Gate job not found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Gate job is not `running` — was never claimed, or was already completed
...
components:
  parameters:
    owner:
      name: owner
      in: path
      required: true
      schema:
        type: string
    repo:
      name: repo
      in: path
      required: true
      schema:
        type: string
...
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    Error:
      type: object
      required: [message]
      description: >
        #97: the one error shape every 4xx carries. `errors` appears on
        validation failures (422) and is a deliberate projection of the
        server's internal validator output — path, message, and a stable
        code — never the validator's raw issue objects, whose shape belongs
        to a library this contract must outlive.
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [path, message, code]
            properties:
              path:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              message:
                type: string
              code:
                type: string
...
    GateJob:
      type: object
      required: [id, repo_id, git_sha, name, image, command, timeout_ms, status, actor_id, created_at]
      description: >
        M4-9a: a queued unit of isolated gate execution. Queue-only — `status`
        transitions queued -> running (via claim) -> a terminal state (via
        complete); no runner attests anything about the isolation it ran
        under here, that lands in M4-9c once a real gate_results row is
        produced from a completed job.
      properties:
        id: { type: string, format: uuid }
        repo_id: { type: string, format: uuid }
        git_sha: { type: string, pattern: "^[0-9a-f]{40}$" }
        name: { type: string }
        image: { type: string }
        command: { type: string }
        timeout_ms: { type: integer }
        status:
          type: string
          enum: [queued, running, succeeded, failed, timed_out, error]
        actor_id: { type: string, format: uuid }
        claimed_by: { type: string, nullable: true }
        exit_code: { type: integer, nullable: true }
        logs: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time, nullable: true }
        finished_at: { type: string, format: date-time, nullable: true }
...